refactor v2 collection reduce fold
